Add new variable bool displayingScreen and new parameter in MenuItem struct, requireUpdateDisplay.
bool displayingScreen is to track wether screen should show menu or fullscreen menu function. While requireUpdateDisplay in the MenuItem Struct is to differentiate whether the function that will be execute need to update the display or not. Mean if the function have something to show on the display, it need to be updated. If the function only have a code to execute without needing to update the display, it don't have to update the display and stay in the current menu.