Refactor ESPNOW code/module
ESPNOW now Send and Received data to sync the data between sender and receiver. When D1 mini first turn on, it will pull data from receiver then turn the wifi back off to save power. The receiver will always send the data every 100ms(the receiver is not using battery. So it is okay to always sending the data)